test: Consolidate and refactor code review integration tests
- Merged multiple code review integration test files into a single comprehensive test suite
- Updated CODEBASE.md to reflect new test file structure
- Improved test coverage for code review API endpoints
- Removed redundant test files and consolidated test cases
- Enhanced error handling and edge case testing for code review operations
